+++ settings.php
@@ -26,7 +26,7 @@
 defined('MOODLE_INTERNAL') || die();
 
 if ($ADMIN->fulltree) {
-    global $CFG;
+    global $CFG, $PAGE;
 
     require_once($CFG->dirroot . '/filter/tabs/filter.php');
 
@@ -38,9 +38,9 @@ if ($ADMIN->fulltree) {
             );
 
     $tabsconfigsoptions = array(
-        \filter_tabs\plugin_config::YUI_TABS => get_string('enableyui', 'filter_tabs', null, true),
-        \filter_tabs\plugin_config::BOOTSTRAP_2_TABS => get_string('enablebootstrap2', 'filter_tabs', null, true),
-        \filter_tabs\plugin_config::BOOTSTRAP_4_TABS => get_string('enablebootstrap4', 'filter_tabs', null, true)
+        \filter_tabs\config::YUI_TABS => get_string('enableyui', 'filter_tabs', null, true),
+        \filter_tabs\config::BOOTSTRAP_2_TABS => get_string('enablebootstrap2', 'filter_tabs', null, true),
+        \filter_tabs\config::BOOTSTRAP_4_TABS => get_string('enablebootstrap4', 'filter_tabs', null, true)
         );
 
     if (($bootstrapversion = \filter_tabs\helper::get_bootstrap_version())) {
@@ -51,7 +51,7 @@ if ($ADMIN->fulltree) {
                 'filter_tabs/enablebootstrap',
                 get_string('selecttabs', 'filter_tabs', null, true),
                 get_string('selecttabs_desc', 'filter_tabs', null, true),
-                '4' === $version ? \filter_tabs\plugin_config::BOOTSTRAP_4_TABS : \filter_tabs\plugin_config::BOOTSTRAP_2_TABS,
+                '4' === $version ? \filter_tabs\config::BOOTSTRAP_4_TABS : \filter_tabs\config::BOOTSTRAP_2_TABS,
                 $tabsconfigsoptions)
             );
 
@@ -59,10 +59,10 @@ if ($ADMIN->fulltree) {
         $suggestedoption = '';
         if ($version !== '4') {
             $suggestedoption .= '<br /><small>' . get_string('suggestedoption', 'filter_tabs', null, true) .
-                                ": [ \"{$tabsconfigsoptions[\filter_tabs\plugin_config::BOOTSTRAP_2_TABS]}\" ]</small>";
+                                ": [ \"{$tabsconfigsoptions[\filter_tabs\config::BOOTSTRAP_2_TABS]}\" ]</small>";
         } else {
             $suggestedoption .= '<br /><small>' . get_string('suggestedoption', 'filter_tabs', null, true) .
-                                ": [ \"{$tabsconfigsoptions[\filter_tabs\plugin_config::BOOTSTRAP_4_TABS]}\" ]</small>";
+                                ": [ \"{$tabsconfigsoptions[\filter_tabs\config::BOOTSTRAP_4_TABS]}\" ]</small>";
         }
 
         // Bootstrap suggestion.
@@ -73,7 +73,9 @@ if ($ADMIN->fulltree) {
             );
 
         // Preview.
-        $filtertabs = new filter_tabs(context_system::instance(), array());
+        $context = context_system::instance();
+        $filtertabs = new filter_tabs($context, array());
+        $filtertabs->setup($PAGE, $context);
         $tabsfilteredtext = $filtertabs->filter('{%:First tab}Some text{%}{%:Second tab}Another text{%}');
         $settings->add(new admin_setting_heading(
                 'filter_tabs_preview_header',